fix types, 64b arguments (breaks things yet)

+++ arch/ppc32/loader/printf.h
@@ -38,8 +38,8 @@ typedef signed char __s8;
 
 typedef unsigned char __u8;
 typedef unsigned short __u16;
-typedef unsigned long __u32;
-typedef long long __u64;
+typedef unsigned int __u32;
+typedef unsigned long long __u64;
 
 typedef __u32 __address;
 typedef __u32 __native;
