+++ arch/ppc64/loader/printf.h
@@ -38,11 +38,11 @@ typedef signed char __s8;
 
 typedef unsigned char __u8;
 typedef unsigned short __u16;
-typedef unsigned long __u32;
-typedef long long __u64;
+typedef unsigned int __u32;
+typedef unsigned long __u64;
 
-typedef __u32 __address;
-typedef __u32 __native;
+typedef __u64 __address;
+typedef __u64 __native;
 
 typedef char *char_ptr;
